+++ src/py42/_internal/clients/alerts.py
@@ -1,5 +1,6 @@
 import json
 
+from py42 import settings
 from py42._internal.compat import str
 from py42.clients import BaseClient
 from py42.clients.util import get_all_pages
@@ -18,34 +19,28 @@ class AlertClient(BaseClient):
         uri = self._uri_prefix.format(u"query-alerts")
         return self._session.post(uri, data=query)
 
-    def get_details(self, alert_ids, tenant_id=None):
+    def get_details(self, alert_ids):
         if not isinstance(alert_ids, (list, tuple)):
             alert_ids = [alert_ids]
-        tenant_id = (
-            tenant_id if tenant_id else self._user_context.get_current_tenant_id()
-        )
+        tenant_id = self._user_context.get_current_tenant_id()
         uri = self._uri_prefix.format(u"query-details")
         data = {u"tenantId": tenant_id, u"alertIds": alert_ids}
         results = self._session.post(uri, data=json.dumps(data))
         return _convert_observation_json_strings_to_objects(results)
 
-    def resolve(self, alert_ids, tenant_id=None, reason=None):
+    def resolve(self, alert_ids, reason=None):
         if not isinstance(alert_ids, (list, tuple)):
             alert_ids = [alert_ids]
-        tenant_id = (
-            tenant_id if tenant_id else self._user_context.get_current_tenant_id()
-        )
-        reason = reason if reason else u""
+        tenant_id = self._user_context.get_current_tenant_id()
+        reason = reason or u""
         uri = self._uri_prefix.format(u"resolve-alert")
         data = {u"tenantId": tenant_id, u"alertIds": alert_ids, u"reason": reason}
         return self._session.post(uri, data=json.dumps(data))
 
-    def reopen(self, alert_ids, tenant_id=None, reason=None):
+    def reopen(self, alert_ids, reason=None):
         if not isinstance(alert_ids, (list, tuple)):
             alert_ids = [alert_ids]
-        tenant_id = (
-            tenant_id if tenant_id else self._user_context.get_current_tenant_id()
-        )
+        tenant_id = self._user_context.get_current_tenant_id()
         uri = self._uri_prefix.format(u"reopen-alert")
         data = {u"tenantId": tenant_id, u"alertIds": alert_ids, u"reason": reason}
         return self._session.post(uri, data=json.dumps(data))
@@ -59,7 +54,29 @@ class AlertClient(BaseClient):
         else:
             return str(query)
 
-    def _get_alert_rules(
+    def get_page(
+        self,
+        tenant_id=None,
+        groups=None,
+        sort_key=None,
+        sort_direction=None,
+        page_num=1,
+        page_size=None,
+    ):
+        # This API expects the first page to start with zero.
+        page_num = page_num - 1
+        page_size = page_size or settings.items_per_page
+        tenant_id = tenant_id or self._user_context.get_current_tenant_id()
+        return self._get_page(
+            tenant_id=tenant_id,
+            groups=groups,
+            sort_key=sort_key,
+            sort_direction=sort_direction,
+            page_num=page_num,
+            page_size=page_size,
+        )
+
+    def _get_page(
         self,
         tenant_id,
         groups=None,
@@ -68,12 +85,17 @@ class AlertClient(BaseClient):
         page_num=None,
         page_size=None,
     ):
+        # This method is meant to called in the `get_all_()` methods and handles paging
+        # through `util.get_all_pages()`. It exists separately than `get_page()` because
+        # of the tenant ID parameter and trying to avoid it.
+
+        # This API expects the first page to start with zero.
+        page_num = page_num - 1
         data = {
             u"tenantId": tenant_id,
             u"groups": groups or [],
             u"groupClause": u"AND",
-            u"pgNum": page_num
-            - 1,  # Minus 1, as this API expects first page to start with zero.
+            u"pgNum": page_num,
             u"pgSize": page_size,
             u"srtKey": sort_key,
             u"srtDirection": sort_direction,
@@ -84,7 +106,7 @@ class AlertClient(BaseClient):
     def get_all_rules(self, sort_key=u"CreatedAt", sort_direction=u"DESC"):
         tenant_id = self._user_context.get_current_tenant_id()
         return get_all_pages(
-            self._get_alert_rules,
+            self._get_page,
             u"ruleMetadata",
             tenant_id=tenant_id,
             groups=None,
@@ -97,7 +119,7 @@ class AlertClient(BaseClient):
     ):
         tenant_id = self._user_context.get_current_tenant_id()
         return get_all_pages(
-            self._get_alert_rules,
+            self._get_page,
             u"ruleMetadata",
             tenant_id=tenant_id,
             groups=[json.loads(str(create_eq_filter_group(u"Name", rule_name)))],
@@ -110,7 +132,7 @@ class AlertClient(BaseClient):
     ):
         tenant_id = self._user_context.get_current_tenant_id()
         results = get_all_pages(
-            self._get_alert_rules,
+            self._get_page,
             u"ruleMetadata",
             tenant_id=tenant_id,
             groups=[