+++ src/py42/clients/detectionlists/departing_employee.py
@@ -1,8 +1,9 @@
 import json
 
 from py42.clients import BaseClient
+from py42.clients.detectionlists import _PAGE_NUM
+from py42.clients.detectionlists import _PAGE_SIZE
 from py42.clients.util import get_all_pages
-from py42.clients.detectionlists import _PAGE_NUM, _PAGE_SIZE
 
 
 class DepartingEmployeeClient(BaseClient):